add menu items to base news fragment
