+++ app/src/main/java/com/sdzshn3/android/news247/Fragments/NewsFragment.java
@@ -5,6 +5,8 @@ import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
@@ -130,6 +132,14 @@ public class NewsFragment extends Fragment implements SwipeRefreshLayout.OnRefre
 
         ItemClickSupport.addTo(newsRecyclerView).setOnItemClickListener(this);
         return rootView;
+    }
+
+    @Override
+    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
+        inflater.inflate(R.menu.main, menu);
+        super.onCreateOptionsMenu(menu, inflater);
+        MenuItem menuItem = menu.findItem(R.id.action_search);
+        menuItem.setVisible(false);
     }
 
     @Override
