+++ app/src/main/java/com/sdzshn3/android/news247/Fragments/TeluguNewsFragment.java
@@ -3,9 +3,6 @@ package com.sdzshn3.android.news247.Fragments;
 import android.content.Context;
 import android.os.Bundle;
 import android.view.LayoutInflater;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
@@ -108,13 +105,5 @@ public class TeluguNewsFragment extends NewsFragment {
 
         ItemClickSupport.addTo(newsRecyclerView).setOnItemClickListener(this);
         return rootView;
-    }
-
-    @Override
-    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-        inflater.inflate(R.menu.main, menu);
-        super.onCreateOptionsMenu(menu, inflater);
-        MenuItem menuItem = menu.findItem(R.id.action_search);
-        menuItem.setVisible(false);
     }
 }
